% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gene_list.R
\name{download_gene_list}
\alias{download_gene_list}
\title{Download OncoKB Cancer Gene List}
\usage{
download_gene_list(path, url = oncokb_dwl_url())
}
\arguments{
\item{path}{A character string with the file path where the downloaded file is to be saved. Tilde-expansion is performed.}

\item{url}{The URL of the resource providing the OncoKB cancer gene list.}
}
\description{
Download OncoKB Cancer Gene List
}
