% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\docType{data}
\name{exported_genes}
\alias{exported_genes}
\title{Exported genes}
\format{
A character vector.
}
\usage{
exported_genes
}
\description{
A character vector of genes for which the retrieval of sequence profiles was
successful and are hence provided with this package.
}
\keyword{datasets}
