% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncokb_genes.R
\name{fetch_oncokb_genes}
\alias{fetch_oncokb_genes}
\title{Fetch current OncoKB genes}
\usage{
fetch_oncokb_genes(file = oncokb_dwl_url())
}
\arguments{
\item{file}{A URL or a file path to the source providing the cancer gene list
file. By default it will automatically download cancerGeneList.tsv from
OncoKB website.}
}
\value{
A character vector of gene names.
}
\description{
\code{\link[=fetch_oncokb_genes]{fetch_oncokb_genes()}} retrieves the current set of OncoKB genes from an OncoKB's
cancer gene list file.
}
\examples{
fetch_oncokb_genes()
}
