% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\docType{data}
\name{missing_genes}
\alias{missing_genes}
\title{Missing genes}
\format{
A character vector.
}
\usage{
missing_genes
}
\description{
A character vector of genes for which the retrieval of sequence profiles was
not successful and are therefore absent.
}
\keyword{datasets}
