% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\docType{data}
\name{oncokb_genes}
\alias{oncokb_genes}
\title{OncoKB genes}
\format{
A character vector.
}
\usage{
oncokb_genes
}
\description{
A character vector of OncoKB genes used as query to retrieve the sequence
profiles bundled with this package.
}
\keyword{datasets}
