% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_path}
\alias{profile_path}
\title{Get the path to a sequence profile}
\usage{
profile_path(file = NULL)
}
\arguments{
\item{file}{Name of file or gene symbol. If \code{NULL}, the profile files will be
listed.}
}
\description{
protean comes bundled with a number of sequence profile files in its
\code{inst/profiles} directory. This function make them easy to access by
returning the local path to them.
}
\examples{
# Retrieve the path to the sequence profile of the TP53 protein
# Using the gene symbol
profile_path("TP53")

# Using the file name
profile_path("TP53.csv.gz")

# List all profile files
profile_path()
}
