% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protean-package.R
\docType{package}
\name{protean-package}
\alias{protean}
\alias{protean-package}
\title{protean: Sequence Profiles of OncoKB Genes}
\description{
A data package of sequence profiles of OncoKB genes. These profiles are obtained via Ensembl's REST API and derived from the pairwise alignment of the human sequence with its orthologs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maialab/protean}
  \item Report bugs at \url{https://github.com/maialab/protean/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Isabel Duarte \email{iduarte.scientist@gmail.com} (\href{https://orcid.org/0000-0003-0060-2936}{ORCID})
  \item Ana-Teresa Maia \email{maia.anateresa@gmail.com} (\href{https://orcid.org/0000-0002-0454-9207}{ORCID})
}

Other contributors:
\itemize{
  \item CINTESIS [copyright holder, funder]
}

}
\keyword{internal}
