% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_profile.R
\name{read_profile}
\alias{read_profile}
\title{Read a sequence profile}
\usage{
read_profile(file = stop("`file` must be specified"))
}
\arguments{
\item{file}{A path to a sequence profile file.}
}
\value{
A \link[tibble:tibble-package]{tibble} of 10 variables:
\describe{
\item{\code{timestamp}}{Date and time of the download from Ensembl.}
\item{\code{human_prot_id}}{Ensembl identifier of the human protein sequence.}
\item{\code{ortho_prot_id}}{Ensembl identifier of the ortholog protein sequence.}
\item{\code{ortho_species}}{Species name of the ortholog sequence.}
\item{\code{human_align_seq}}{In the context of pairwise alignment between the
human sequence and one of its orthologs, this is the aligned human sequence.}
\item{\code{ortho_align_seq}}{In the context of pairwise alignment between the
human sequence and one of its orthologs, this is the aligned ortholog
sequence.}
\item{\code{human_ortho_perc_id}}{Percentage of the human sequence matching the
sequence of the ortholog.}
\item{\code{ortho_human_perc_id}}{Percentage of the orthologous sequence matching
the human sequence.}
\item{\code{human_profile_id}}{Human protein sequence.}
\item{\code{ortho_profile_seq}}{Orthologous sequence stripped off of the alignment
positions which correspond to gaps in the human sequence.}
}
}
\description{
Read a sequence profile
}
\examples{
read_profile(profile_path("TP53"))
}
