% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_profile.R
\name{write_profile}
\alias{write_profile}
\title{Get and save a sequence profile to a file}
\usage{
write_profile(
  symbol,
  file = stop("`file` must be specified"),
  overwrite = FALSE,
  warning = TRUE
)
}
\arguments{
\item{symbol}{A gene symbol to be used as query.}

\item{file}{A path to a file for exporting the sequence profile.}

\item{overwrite}{Whether to overwrite the file if it already exists.}

\item{warning}{Whether to print a warning if, for some reason, the sequence
profile for the queried gene symbol was not retrievable.}
}
\value{
This function is run for its side effect, i.e. writing the FASTA
file.
}
\description{
This function fetches a sequence profile via Ensembl's REST API and saves it
to a file.
}
\keyword{internal}
