% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_fasta.R
\name{write_to_fasta}
\alias{write_to_fasta}
\title{Write sequence profile as FASTA}
\usage{
write_to_fasta(seq_profile, file = stop("`file` must be specified"))
}
\arguments{
\item{seq_profile}{A sequence profile, i.e. a data frame obtained with either
\code{\link[=get_profile]{get_profile()}} or \code{\link[=read_profile]{read_profile()}}.}

\item{file}{A path to a file for exporting the sequence profile.}
}
\value{
This function is run for its side effect, i.e. writing the FASTA
file.
}
\description{
Write sequence profile as FASTA
}
\keyword{internal}
